thermal: Use separate shutdown thresholds for the thermal core and thermal-engine.
Currently, the thermal core and thermal-engine shutdown threshold are the same.
Need to push the thermal core shutdown threshold to 2 degree C higher for cushion.
Because the thermal core won't provide the shutdown reason but the thermal-engine does,
So we keep the both of them as a backup if the normal mitigation path does not work (thermal-hal).

Bug: 128935264
Test: Verified thermal_core shutdown via emul temp.
[  213.956392] c3   6953 thermal thermal_zone34: critical temperature reached(59 C),shutting down
......
[  116.054398] c5   4743 thermal thermal_zone34: critical temperature
reached(62 C),shutting down

Change-Id: I30593028c3b7eeef35b290c5b009bedeec7f8664
Signed-off-by: davidchao <davidchao@google.com>