mfd: adnc: play hotword after start ambientaudio leads crash
Event subscribe and unsusbscribe functions has specific register
read/write, update block sequence and they are supposed to execute
automically. This was not ensured and this was leading to FW error
update block as the sequence were getting mixed. This is fixed by
introducing mutex to protect the register sequences.

Testing Done:
 Verfied the mentioned use case in JIRA multiple times

Bug: 130402903
Change-Id: I4215dbd3d37cdc9647e2e2bb9d10dc6f075b81c2
Signed-off-by: Aravind Retnakaran <aravind.retnakaran@knowles.com>