arm64/dts: floral: abc-pcie under pcie_rc1
According to the change commit d4001297951b ("pci/of: expand match PCI
devices to OF nodes") we need to use a pci-ids to bind abc-pcie device
back to an OF device tree node. Previously we borrow an mhi_0, which
does not work now since the pci-id is different from that of abc-pcie.

Bug: 129091813
Change-Id: I55c0bd54a578ae4d7a6a96cde94efa46ae339dd8
Signed-off-by: Ahbong Chang <cwahbong@google.com>