char: oscar: allow mapping ab-dram buffer multiple times
Allow the same Airbrush DRAM buffer to be mapped to the TPU multiple
times.

Add a new AB-DRAM unmap ioctl that specifies which mapping is to be
unmapped, identified by TPU device address.

Mark the existing AB-DRAM unmap ioctl and supporting metadata as
deprecated, to be removed after the TPU runtime is in sync with the
new kernel driver code in the coral development branch.

AB-DRAM buffer metadata reference count is a count of active mappings
plus one for each temporary reference held by code accessing the
metadata.  Release AB-DRAM buffer metadata when the last mapping is
unmapped.  Also release on device close or reset as before, but no
longer unmap the buffer on release, as we no longer keep track of all
mappings for the buffer, and this isn't necessary -- the MMU mappings
are about to be reset.  Furthermore, AB-DRAM may be powered down or in
retention, or the PCIe link may be unavailable, when the device
close/reset/HAL process exit occurs, such that extended level-1
subtables are not available.

Bug: 129156996
Bug: 129019555
Change-Id: I55f734f8268325ee7763f8404b845b828ffca37b
Signed-off-by: Todd Poynor <toddpoynor@google.com>