abc: tmu: trip temperature could be reset to its value in OF
For emulated temperature (emul_temp in sysfs) it could be disabled by
set to 0. We also need way for disable customized trip temperatue and
trip hysterisis.

There is only set_trip_temp() but not set_hyst_temp() for OF thermal
zone, so we let set 0 to trip temperature to reset the value of both
the temperature and hysterisis of that trip point.

Bug: 129685235
Change-Id: Ic6d8716b00f82b8c73c64cd2504ad9f36b8aafe8
Signed-off-by: Ahbong Chang <cwahbong@google.com>