char: oscar: remove obsolete oscar interfaces
Remove oscar driver alternate interfaces for allocating, performing
DMA from/to, and deallocating Airbrush DRAM.  The oscar runtime no
longer uses these ioctls; only the map and unmap ioctls are specific to
the TPU and still in use.  The old-style unmap ioctl will be removed
when the runtime is updated to switch to it.

Remove the clock gating ioctl that was never implemented and is not
called by the runtime.

Remove the module parameter for "bypass top level" / "only GCB
instantiated", a holdover from previous chips that is no longer in use.

Bug: 129156996
Change-Id: Iec198d2771b56db8530fc53683c00eb920a853d0
Signed-off-by: Todd Poynor <toddpoynor@google.com>