misc: cs40l2x: Retract suspend blocker following playback error
If cs40l2x_vibe_start_worker exits early in the event of an error,
a suspend blocker remains present yet the subsequent stop-playback
event (which normally removes the suspend blocker) doesn't arrive.

In this case, the device cannot enter runtime suspend. This patch
removes the suspend blocker if cs40l2x_vibe_start_worker does not
run to completion as a precaution to avoid such a condition.

Bug: 129998555
Test: Verify haptics still work.
Change-Id: I4ea89937031965fdf4f002bc61347fa8218847d2
Signed-off-by: Jeff LaBundy <jeff.labundy@cirrus.com>
Signed-off-by: Harpreet "Eli" Sangha <eliptus@google.com>