misc: ab: pmic: Shift 8 ms delay from ldo5 to ldo4
In theory, it doesn't matter whether this 8 ms delay is at the time
of disabling LDO5 or disabling LDO4.  But in reality, VDDP has been
found to cause VDDIO to plateau.

Therefore, in this patch, the 8 ms delay is inserted between turning
off VDDP and VDDL to meet HW shutdown sequence requirement.

Test: repeated suspend stress test
Bug: 124518678
Change-Id: I0d2ff40810e5c2df23826a646c2452e7603a8b7e
Signed-off-by: Cheng Gu <gucheng@google.com>