ab: sm: Check EL2 context state before performing enter/exit EL2
Subrscribe to Qualcomm's Subsystem framework to get notifications when
the system is about to enter the faceauth secure world (EL2) and when
it's finished exiting.

Entering and exiting the faceauth secure world is controlled by the
/dev/subsys_faceauth file.

Ensure that the system is out of the secure world before performing the
enter/exit EL2 operations. This is due to the fact that the PCIe link is
mapped to the EL2 context when EL2 is enabled and can lead to crashes
when AB SM interacts with it.

Bug: 129341384

Change-Id: I5e36c7e401f3d0e09db22000152e96f287b4f71f
Signed-off-by: Jawad AlDhalemi <jaldhalemi@google.com>