tcpm: Send softreset instead of reject
From Deterministic PD Compliance MOI
TD.PD.SNK.E9. GetSinkCap in Place of Accept
Applicable to:
- PHY: BMC, FSK
- UUT: UFP
Assertions Tested:
6.3.13#1, 6.3.13#5, 6.7.1#2, 6.7.1#5
Description:
As Provider, the Tester intentionally sends a GetSinkCap message in place of Accept message and
verifies the UUT will send a SoftReset and recover from the error.
Deterministic PD Compliance MOI
Sink Tests | 63 of 116
Steps:
a) Run PROC.PD.E1 Bring-up according to the UUT role.
b) The Tester cycles VBus.
c) The Tester sends a Source Capabilities message to the UUT.
d) Upon receipt of a Request message from the UUT, the Tester replies with a GoodCRC message.
e) The Tester sends a GetSinkCap message to the UUT.
f) If a SoftReset is not received within 15 ms after the GetSinkCap EOP was sent, the test fails.
g) If a SoftReset is received timely, the Tester replies with an Accept message.
h) The Tester sends Source Capabilities message to the UUT repeatedly until nCapsCount reached
or a GoodCRC is received. If nCapsCount reached, the test fails.
i) If a Request is not received timely within 30 ms after the GoodCRC EOP corresponding to
Source Capabilities message was received, the test fails.

Bug: 129797124
Bug: 149187796
Change-Id: I75c62132e2d9088a47877b0ca20d194c7e0b6579
Signed-off-by: Badhri Jagan Sridharan <badhri@google.com>
Signed-off-by: rickyniu <rickyniu@google.com>