mfd: adnc: Disable Wakeup Event
+Host Driver currently uses Wakeup event to know that the
chip is woken up from optimal/sleep. Since Wakeup can be triggered
by a regmap_read from ISR itself, it is not safe to wait for wakeup
event in PM Wake-up logic. Removed subscription for Wakeup event and
added polling in PM Wakeup logic to wait for chip to wake up
for specified amount of time.

Bug: 132802056
Test: soundtrigger function works
Change-Id: I7c2b4d6c81be974b1fd99406006de2cd85a1eb82
Signed-off-by: Shanmugam Vasudevan <Shanmugam.Vasudevan@knowles.com>