upstream: filemap: don't unlock null page in FGP_FOR_MMAP case
We noticed a panic happening in production with the filemap fault pages
because we were unlocking a NULL page.  If add_to_page_cache() fails
then we'll have a NULL page, so fix this check to only unlock if we
have a valid page.

minchan: this patch was squeezed by [1] on upstream kernel.
[1] a75d4c3337727, filemap: kill page_cache_read usage in filemap_fault

Bug: 129765516
Change-Id: Ifd464789fc14dd179d421c78142086450f62ad8f
Signed-off-by: Josef Bacik <josef@toxicpanda.com>
Reviewed-by: Rik van Riel <riel@surriel.com>
Signed-off-by: Minchan Kim <minchan@google.com>