mfd: adnc: Add ioctl to let HAL know the FW state
HAL need info regarding the state of FW in the event of crash in
user space and recovery. It would like to know if any route is active or
any plugins are loaded. In that case it would try to reset the firmware
and re-run the existing use case (if needed).

Implemented two new IOCTLs to get the FW state and to do FW reset

Bug: 129505523
Test: ASR could recovery
Change-Id: Id159bf35e434bd029f04297cf680a52515fc99ae
Signed-off-by: Aravind Retnakaran <aravind.retnakaran@knowles.com>