ab: ddr: Momentarily turn on LDO2 rail to configure MR registers
The VDDQ rail is off during the Self-Refresh. But in Suspend,
we configure the MR registers to set 1866MHz timing parameters.
As VDDQ(LDO2) is off, the timing parameters will not be set and
the ddr init & train during suspend boot may not be proper.
Hence turn on LDO2 rail momentarily to configure MR registers.

Bug: 130115644

Change-Id: I28f89196e0d07c26c496dfaac00bfe974f4607be
Signed-off-by: Shaik Ameer Basha <shaik.ameer@samsung.com>
Signed-off-by: Raman Banka <raman.k2@samsung.com>