[REVERT ME] kernel: printk: Limit amount of logs outputted to console
console_unlock() will attempt to flush all logs to console in one
go. If there is a large burst of logs, it will potentially hog a
cpu for seconds if the caller of console_unlock() is in interrupt
disabled context. Limit the amount of logs being outputted to console
to 75 lines per call if serial console is enabled. The next call
to printk will take over and continue flushing logs to serial console.

This is a temporary workaround to prevent watchdog resets because
current builds have a large amount of selinux denials during boot.
This workaround should be reverted once the amount of logs are at
a reasonable state.

Bug: 125840203
Change-Id: Ib60f4937769c98d1279ec2c0f339556d76b2ddc5
Signed-off-by: Jonglin Lee <jonglin@google.com>