pd_engine: Update phy params when setting roles
When initially calling to pd_phy_open after attaching to a PD sink, the
pdphy_params are not set correctly to SRC/DFP. A wrong PD message
header with Power-Role/Data-Role set to Sink/UFP in the first GoodCRC
from the device will be sent to the port partner. Fix this by correctly
setting the Power-Role/Data-Role in every calling to tcpm_set_roles.

Bug: 120644802
Bug: 131162105
Change-Id: Icb47a4ae656450b21f19ce42dfb8d3b60b618f6b
Signed-off-by: Kyle Tso <kyletso@google.com>