misc: ab: sm: Combine IPU/TPU block_set_state methods
Created PMU methods that sleep and resume IPU and TPU together.
This allows us to wait for IPU and TPU block sleep/resume in
parallel instead of in serial.

Split IPU and TPU clock rate methods into start and finish
components. The start method updates the pll, and finish
waits for the pll to lock. This allows us to wait for both
IPU and TPU pll lock times in parallel instead of in serial.

Bug: 130236049
Bug: 130958250
Change-Id: I8a9ebf02e4c869f293e602693c1b34831bc9b492
Signed-off-by: Shaun Kelsey <shaunkelsey@google.com>