display: schedule a work to handle HBM commands from upper layer
From screen off to on, there is a little bit of time needed for
MDP/DSI host to get initialized properly.

This change is made to schedule a work to handle HBM commands.
In this way, the driver will get enough time to be ready before
it processes the coming command.

BUG: 130840859

Change-Id: I7ed8601bd4702d88d1534dd7c10c71acfeebd483
Signed-off-by: chester_sun <Chester_Sun@compal.com>