smb5-lib: Update SW_ICL_MAX_VOTER in typec_src_insertion
If smblib_handle_rp_change is skipped in typec_state_change_irq_handler
because sink_src_mode has not been updated yet, SW_ICL_MAX_VOTER will
not vote for BC1.2 in the IRQ handler. Device will not charging if the
vote was previously enabled with 0 and APSD is done before the Type-C
attach IRQ.

Update the voter in typec_src_insertion in case that the vote is skipped
in typec_state_change_irq_handler.

Bug: 131891332
Change-Id: I4292e1ef7d334b6a3a1479d6a024511861ae0148
Signed-off-by: Kyle Tso <kyletso@google.com>