pd_engine: Re-design the PD reset in shutdown callback
tcpm_port_reset will set the CC to Open and leave the CC pins in
floating which causes the CC connection fail if tcpm_port_reset is
called in shutdown process.

In order to reset the PD connection in shutdown process, silently send a
hard_reset to port partner and disable PD capability in the upcoming
re-connection.

Bug: 132154352
Change-Id: I272e5bcb82670e9e06380530b923dc55af7844d2
Signed-off-by: Kyle Tso <kyletso@google.com>