bpf: relax inode permission check for retrieving bpf program
For iptable module to load a bpf program from a pinned location, it
only retrieve a loaded program and cannot change the program content so
requiring a write permission for it might not be necessary.
Also when adding or removing an unrelated iptable rule, it might need to
flush and reload the xt_bpf related rules as well and triggers the inode
permission check. It might be better to remove the write premission
check for the inode so we won't need to grant write access to all the
processes that flush and restore iptables rules.

Change-Id: Ie622e772ca38ab92996a60f661705bb9277f390f
Git-commit: e547ff3f803e779a3898f1f48447b29f43c54085
Git-repo: https://android.googlesource.com/kernel/common/
Signed-off-by: Chenbo Feng <fengc@google.com>
Signed-off-by: Sharath Chandra Vurukala <sharathv@codeaurora.org>