ab: clkout: reset the counters after freq measurement
The current h/w implementation of the clkout counter
is such that, it only allows clearing of the counters when
clkout0/1 clocks are supplied (enabled and non-zero).

In case clkout0/1 is gated or supplied with 0Hz clock, the
reset control of the clkout counter will not able to clear the
previous counter values (doesn't reset the counters to zero).
This will provide wrong counter values when trying to measure
a clock with 0 Hz frequency.

To fix the above issue, reset the clock counters once the
frequency measurement is completed.

Bug: 132443661

Change-Id: Ic96df70185c0258b841594ab6549dba4f72c1310
Signed-off-by: Shaik Ameer Basha <shaik.ameer@samsung.com>
Signed-off-by: Thiagu Ramalingam <thiagu.r@samsung.com>
Signed-off-by: Shaun Kelsey <shaunkelsey@google.com>