display: add ESD recovery mechanism
This recovery mechanism is implemented to check the panel status
through reading the status register from the panel every time
the display state is turned to screen on.

In order to ensure the panel can recover form ESD failure, it will
be reset fully if it were in a bad state.

BUG: 127727224

Change-Id: I847ce0c622b2b0ca54fcd46d8adec9da1db11790
Signed-off-by: chester_sun <Chester_Sun@compal.com>