arm64/dts: floral: reduce transfer time during 60hz modes
Mismatch between transfer times in 60hz and 90hz mode is leading to
tearing (observed as a gray line in the middle of the screen).
Matching the transfer times between both modes, can address this issue
at slight power cost due to increased clocks. Use this as a workaround
until we identify a cleaner solution.

Bug: 132240539
Change-Id: I7aa08842a306b80ce3583f5a06dc07cbdd4086d9
Signed-off-by: Adrian Salido <salidoa@google.com>