Makefile: append BUILD_NUMBER to version string when defined
In current implementation the Linux version string that is attached to
the kernel linux image as well as the all kernel modules includes the
sha1 of the main git project for the kernel.

This Commit adds the BUILD_NUMBER env variable with an 'ab' prefix to
the Linux version string. This maps to both the Linux Image and all
kernel modules version strings.

Bug: 138639911
Test: Build the kernel and test the version string
Change-Id: I621ffe5dd119b7db042528087355850ce8cf5565
Signed-off-by: Robin Peng <robinpeng@google.com>
(cherry picked from commit 3b1163c066ef707afb9c2c7f0e69708bbcd079eb)