staging: s2mpg01: Use resetb falling edge only
Only requests resetb falling edge interrupts to avoid confusion.

When resetb interrupt triggers, treat as s2mpg01 reset and poll
for coming back from reset.

Removes duplicate calls to s2mpg01_notify_fail_all().  Such notification
is only generated by resetb interrupt handler with this patch.

Downside of this patch is that the signal of resetb rising edge
is not reflected immediately, but depends on the polling interval.
One example is it can take a long time to clear TSD, due to hysteresis.
This should be fine for other reset reasons such as PON, WATCHDOG and
TH_TRIP, as they are either in user's control or can get out of
reset quickly.

Bug: 135209696
Change-Id: I4258718eefe5b5a24de7f4563378ac864097de29
Signed-off-by: Cheng Gu <gucheng@google.com>