binder: instrumentation to catch corruption for b/135345566
There is a data corruption issue on floral which,
when it occurs, always corrupts the alloc->pages
array starting at offset 0x1a0. The problem only
manifests when that process exits and the pages
array is being cleaned up -- so the crash can
occur a long time after the corruption occurs.
This patch checks for corruption often to force
the crash to be closer to when the corruption
occurred.

Bug: 135345566
Test: tested on flame (safe if no corruption)
Change-Id: Ib7140a0ef3312226265a08426a9650a92ff8c106
Signed-off-by: Todd Kjos <tkjos@google.com>