tcpm: Prevent log overflow by removing old entries
TCPM logs overflow once the logbuffer is full. Clear old
entries and allow logging the newer ones as the newer would
be more relevant to the issue being debugged.

Also, do not reset the logbuffer tail as end users might
take back to back bugreports which would result in an
empty buffer.

Bug: 136668849
Change-Id: I101e9db6c60958194bab786951d974a2b45f9a4d
Signed-off-by: Badhri Jagan Sridharan <badhri@google.com>