adnc: mfd: Fix for deadlock in pm_suspend/resume
+Ensure power management functions use right
regmap on power up and power down.

+Waiting for update_block_lock in pm_suspend path
causes deadlock since there is possibility that the
thread that is holding the update_block_lock
is waiting for another thread (pm_resume) that in turn
is waiting for pm_suspend to finish. Remove the
update_block_lock in the function that waits for
update block status to be cleared if the regmap used
is regmap_no_pm. Since separate regmap is used to
read the register, there is no need to use the
update_block_lock in that case.

+Remove unused functions in power management module

Testing Done:
Tahoe tests passed 25/25.

Bug:137195743
Test: video call + hotword + ambient + get_pwr_stat stress test

Change-Id: Iea2dedcbd64d313495ec8385bf59f327594a7d69
Signed-off-by: Shanmugam Vasudevan <Shanmugam.Vasudevan@knowles.com>
Signed-off-by: rogerfang <rogerfang@google.com>