smb-lib: continue overheat mitigation until cable unplugged
When the overheat mitigation is running, and the temperature
has dropped below the threshold, sometimes the overheat mitigation
would end without cable unplugged.

This is because wrong cc attachment status was obtained before it
was ready to read Type-C cc status.
This error determination can be resolved by checking the debounce
done status to ensure that the Type-C cc status is ready.

Bug: 137164444
Change-Id: I90ef781bb9db4411436907e361261a8c8c684261
Signed-off-by: Jack Wu <wjack@google.com>