misc: mnh: Fix race condition in legacy fw transfer
The legacy firmware transfer routine looks for FW_IMAGE_DOWNLOADING
flag to determine whether to wait for download complete + to unmap
memory.  If a DMA interrupt comes very quickly, this flag will be
overwritten and it's possible that the routine reports an error
while the transfer is actually successful.

This patch fixes the issue by checking any flag other than
FW_IMAGE_NONE.

Bug: 111524987
Change-Id: I6b84e420f19e5d220da99b2938a5c27f8cdba298
Signed-off-by: Cheng Gu <gucheng@google.com>