misc: ipu: track client refs and reset req separately
Previously there was a JQS state variable, runtime_requested to
track whether the client had requested the JQS through rpm.  This
proved to be unreliable as the rpm hooks are also invoked as part
of dpm.  The runtime_requested variable was also overloaded to
prevent an order suspend in the period between when the reset was
requestd and when the reset had completed.

This change eliminates runtime_requested and replaces it with a
client reference count and a reset_in_progress flag.  The client
reference count is incremented when the client is opened and
decremented on close.  This reference count is indepenent of the
rpm hooks so it should not be affected by dpm.  The
reset_in_progress flag is set when the reset is requested and is
cleared when the reset request has been completed.

Bug: 138236065
Test: atest :airbrush-platform-presubmit + E2E tests of Bokeh and
Ultradepth
Change-Id: I719263ae43732460cba03541a452236b5be6f4a4
Signed-off-by: Adam Hampson <ahampson@google.com>