usb: dwc3: Queue sm_work only on sm_usb_wq
If the sm_work is queued on different workqueues at
the same time, both the work items can be executed at
the same time may result in the state machine run for
the same event and the same state twice which is
unwanted. Fix this by queueing sm_work only on the
dedicated workqueue.

Bug: 139730035
Change-Id: Idfd8ad9183018d35f04d6be284e24d5bb95bec11
Signed-off-by: Sriharsha Allenki <sallenki@codeaurora.org>
Signed-off-by: Puma Hsu <pumahsu@google.com>