misc: airbrush: go to final freq for ipu states
When we move from off state/ tpu state to IPU only state we usually
first change to lowest IPU freq and then to the final freq. a reset bug
showed that the multiple changes can result in an error. The update
makes sure that we change to the final IPU rate directly.

Bug:138942419

Change-Id: I25b388c9ad23bc6fa78dd7027c363d2d21478c59
Signed-off-by: isca <isca@google.com>