misc: ipu: stop using PM infrasructure
Previously we were using rpm to track when the JQS should be shut down.
this was proven to be unreliable as the rpm hooks are invoked when the
phone goes to suspend even if the ipu is still active in the background.

This change moves the IPU logic to use the existing client_count
reference count to control when the JQS should be shut down.

Bug:139322098

Change-Id: I6fb4a8531271c829a7c596d7579988a79aa8ad42
Signed-off-by: isca <isca@google.com>