misc: airbrush: go to higher ipu state when load JQS on transition
When we move from a suspend state to a TPU only state we need to load
the JQS to IPU. Until now this was done in the lowest IPU power state
but this seems to cause timeouts and watchdog errors. This change
updates the power state to a higher one to avoid the errors. since the
change is for a very short time this shouldn't have a large effect on
power consumption.

Bug:139640847


Change-Id: I1562fa8f72fc71378ff04e18f58f8a2ccd451bde
Signed-off-by: isca <isca@google.com>