smb5: Vote 0 on USB_ICL in the beginning when PD is set to active
When the Source establishes PD connection, smb5 will vote 100mA on
USB_ICL. During PR_SWAP while the Source transitioning to Sink, Vbus
from the new Source may come earlier and the new Sink starts charging
because of the 100mA vote. However, the PD state machine will vote 0 in
the early state of power negotiation which causes the charging states to
be flaky.

Fix this by changing the vote to 0mA so that after the PD is set to
active, the Sink will start charging only if the explicit contract is
established.

This change also fixes TDA 2.3.3.1 and 2.3.3.2 of the PD Compliance
Test. In the test items, the tester sends 0A PDO to DUT and expects DUT
to suspend the current. Before DUT handles the 0A contract, it will draw
the current upto the voted value and these items fail if the value is
higher than 0A.

Bug: 129744695
Bug: 139839485
Change-Id: Id19b749028b84fb6e76272b7cf8cc784f1841b69
Signed-off-by: Kyle Tso <kyletso@google.com>