google_charger: Keep waking-up while PPS is active
PPS adapter timeouts if sink doesn't re-send the Request in 12 - 15
seconds. When system is resuming from suspend, Qualcomm PD phy will
return error in the first few seconds when we try to ping the adapter.
To prevent the timeout, hold a wakelock when PPS is active to re-send
the Request in time.

Remove this wakeup source once we fix the issue in the PD phy.

Bug: 147568739
Bug: 136520031
Bug: 138700648
Change-Id: Iea833e5f76ee8af694a8f842354a84fee1bca1b6
Signed-off-by: Kyle Tso <kyletso@google.com>
Signed-off-by: Puma Hsu <pumahsu@google.com>