google_charger: Keep waking-up while PPS is active
PPS adapter timeouts if sink doesn't re-send the Request in 12 - 15
seconds. When system is resuming from suspend, Qualcomm PD phy will
return error in the first few seconds when we try to ping the adapter.
To prevent the timeout, hold a wakelock when PPS is active to re-send
the Request in time.

Remove this wakeup source once we fix the issue in the PD phy.

Bug: 149055293
Change-Id: Iea833e5f76ee8af694a8f842354a84fee1bca1b6
Signed-off-by: Kyle Tso <kyletso@google.com>
(cherry picked from commit 0f3b571dcd94802e02f798d788a4b1181f8c1573)
(fix the conflict and change the method to register the wakeup source)