google_charger: Remove PPS profile when battery is full
We don't need PPS charging when battery is full. Switch the profile to
Fixed PDOs and remove PPS profile in this situation. And recover it
when charging source is removed.

Also fix the typo of the function name.

Bug: 149055293
Change-Id: I7600726a4241039f27ca93528ddd31f2ff56ccbf
Signed-off-by: Kyle Tso <kyletso@google.com>
(cherry picked from commit 5e17a8920acce6291f57c9662959bcd05f3186f8)