misc:faceauth: Reducing hypx blocking time.
When we are done with faceauth, we need to wipe all the
sensitive info and code itself. It will take some time for the
DSP to finish that. Previously the wait is purely done in hypx.
This will block all the SCM call during the waiting due to the
fact that they shared the same scm_mutex and suffering from lock
contension. In this case, we move the wait into kernel to achieve
better utilization. By moving the wait into kernel, we can avoid
long preempt-disable call so that it won't block high-priority
functions.

Bug: 138555694
Test: Boot P19, atest+e2e
Change-Id: I05627be5ffeace2948ce2ee84227d7f34dfd21c5
Signed-off-by: Wei Miao <miaowei@google.com>