ab: ddr: phy reinit and train during dll lock failures
As part of self-refresh exit, to reduce the transition time,
the current code performs the dll locking after the PHY input
clock is moved to PLL. This was done instead of the phy reinit
and train sequences.

As we have seen dll lock failures during the self-refresh
exit, perform the actual phy reinit and train seqeunces to
recover from the dll lock failures.

Instead of replacing the dll locking logic from self-refresh
exit, this patch performs the phy reinit and train only
during dll lock failures.

Bug: 141129421
Test: atest :airbrush-platform-presubmit
      E2E faceauth test
      E2E camera test (front + back)


Change-Id: If244cfadfce499d86139405742d40c9069e5c671
Signed-off-by: Shaik Ameer Basha <shaik.ameer@samsung.com>
Signed-off-by: Thiagu Ramalingam <thiagu.r@samsung.com>