misc: ipu: updated queue file callbacks
originally the IPU queue fd closes the queue when the flush function for
the file is called. b/138963974 showed the fact that the flush function
will be called if a copy of the file is closed this leads to errors
since the original fd can still be open. this logic updates make sure
that the queue is only closed when the original fd is closed.

Bug:138963974

Change-Id: I17fcc82d8d76d8b368de7bc38d86acdc66438c75
Signed-off-by: isca <isca@google.com>