regulator: tps: Add support for TPS regulator
This driver adds support for the TPS fixed voltage regulators
wired to a set of GPIOs and when the GPIOs are needed to be flipped
on/off in a specific sequence to enable/disable regulator.
The driver is needed when more than one gpio pin is needed to control
the regulator. Optionally, a delay in microseconds can also be
specified after enabling/disabling each gpio.

Bug: 141281367
Change-Id: I5bcb9899fb562be64145b7c5306a652e300b44db
Signed-off-by: rickyniu <rickyniu@google.com>
Signed-off-by: Badhri Jagan Sridharan <badhri@google.com>
(cherry picked from commit 21135cc4ba80838766625d285dce2a62981332c7)