build.config.kasan: explicitly use KASAN_OUTLINE
Due to a bug in lib/Kconfig.kasan, we have been using KASAN_OUTLINE
implicitly for KASAN image builds.

In preparation for fixing the bug in lib/Kconfig.kasan, we now need
to explicitly use KASAN_OUTLINE in the KASAN build config.

This change originates from:
commit 59311599b590 ("kasan: fix instrumentation modes choice config")
Author: Andrey Konovalov <andreyknvl@google.com>

but it needed to be split into build.config change and Kconfig change.

Change-Id: Ie3b9816343fc6a00e4b21f6ef4aace03d46113bc
Signed-off-by: Petri Gynther <pgynther@google.com>