build.config.{bluecross,bonito}_kasan: explicitly use KASAN_OUTLINE
Due to a bug in lib/Kconfig.kasan, we have been using KASAN_OUTLINE
implicitly for KASAN image builds.

In preparation for fixing the bug in lib/Kconfig.kasan, we now need
to explicitly use KASAN_OUTLINE in the KASAN build config.

This change originates from:
commit 59311599b590 ("kasan: fix instrumentation modes choice config")
Author: Andrey Konovalov <andreyknvl@google.com>

but it needed to be split into build.config change and Kconfig change.

Change-Id: Iab25168de4e9586ea3cb90567f719a255fb2aca2
Signed-off-by: Petri Gynther <pgynther@google.com>