GKI: drivers: soc: qcom: move ownership of PIL-related objects
The peripheral-loader (PIL) driver provides a function, a pointer, and
tracepoints used by the susbsytem_restart driver. So, by default, the
latter should be a module for the former to be modularized. However, the
susbsystem_restart driver has many complex dependencies making this hard
to do.

Since the PIL-owned objects used by subsytem_restart are relatively
easy to move out of it, reverse the dependency by moving them to
subsytem_restart, which is the only piece of code needing it.

Bug: 144523239
Test: modules load, same devices probe on lito
Change-Id: I7f5912f8df46b4079a42edf13297d10ca8a98b02
Signed-off-by: Quentin Perret <qperret@google.com>