input: touchscreen: touch_offload driver
The touch_offload driver augments traditional touchscreen drivers with
the ability to capture and export heatmap and coordinate touch data for
processing in user space. The driver's client can query the touch data
available, configure which content is captured, read queued data as it
is captured, and report detected touch events back to the driver.

Bug: 139955910
Signed-off-by: Steve Pfetsch <spfetsch@google.com>
Change-Id: Ifa2f003f281bfdf6f289ee3b15cd1ee47220f0ec