GKI: asoc: rt5514: move dependencies between rt5514 and rt5514-spi (part deux)
This fix moves rt5514_set_gpio, a new circular dependency from rt5514
into rt5514-spi. And once again supports rt5514 as a module, and
permits rt5514-spi to be a module or built-in.

History

A regression that introduced a circular dependency in
commit bbfeecb8728e6bb5c867feb4ca2a43f235c5279d
("asoc: rt5514: Update driver from vendor") also broke the ability to
build rt5514 as a module, and rt5514-spi as built-in as
emboddied in the Kconfig.  A fix in
commit a4fb8046cf2d1e3f6914dbf1f9457dd984ecfa85
("GKI: asoc: rt5514: remove dependencies between rt5514 and rt5514-spi")
allowed that configuration to continue and resolved that circular
dependency.

A new regression that again introduced a circular dependency in
commit 50462bff1394ad8ccbe14736fa35d2f728aa175b
("2 Host SPI control for RT5514") continues down the path to
break both supported configurations of rt5514 as a module.

Signed-off-by: Mark Salyzyn <salyzyn@google.com>
Bug: 145324997
Bug: 143146942
Bug: 144327705
Bug: 143510877
Test: build with rt5514 as module, rt5514-spi as module or built-in
Test: build with rt5514 as disabled, rt5514-spi as module or built-in
Change-Id: I1f279d10e3de61d00ad7ba1b32ea86f482de08ad