GKI: drivers: thermal: Modularize driver for qcom,msm-limits-cdsp
When built-in, this driver creates a msm_isense_cdsp folder under
/sys/module to expose some data to userspace. But when built as a
module, the /sys/module/msm_isense_cdsp folder gets created upfront by
the module subsystem. Unfortunately, the driver tries to create the
folder unconditonally, which fails in the module case, and causes the
whole driver probing to fail.

To avoid this issue, re-factor the kobject creation code by checking the
existence of a matching kobject under the module_kset, and use it
directly if found instead of creating it from scratch.

To complete the modularization, add the missing licence as well as a
seemingly obvious exit path.

Bug: 144405533
Test: compiled, booted on lito, same devices probe
Change-Id: Iedc837b2f4a28ccc12b139a108633fae1d023994
Signed-off-by: Quentin Perret <qperret@google.com>