GKI: scripts: stop if export is __init or __exit
Strip problematic __init or __exit tags from exported functions, but
stop with script failure so that the human can intercede and run a
build to determine the side effects and inspect if other changes are
required.  The script can be run again with the _original_ build log
file to complete the operation.

Add a check to confirm that the git files that are changed match the
expected files to be adjusted and warn if they are different.  Submit
the whole overlapping set as a means of allowing continuation of the
script.

Manual Test involves generating a false build log, for example:

 :: undefined reference to `cam_cdm_intf_init_module'
 :: undefined reference to `cam_cdm_intf_exit_module'

and letting it modify sources that have __init or __exit tags to
prove out the functionality.

  $ add_EXPORT_SYMBOL_GPL < build_output_file
  Reading symbols
  Finding symbols in the current tree
  Editing the files that contain the symbols
  INFO: export cam_cdm_intf_exit_module in drivers/cam_cdm/cam_cdm_intf.c
  int __init cam_cdm_intf_init_module(void) -> int cam_cdm_intf_init_module(void)
  ERROR: scrubbing __init from function in drivers/cam_cdm/cam_cdm_intf.c
         Compile check, possibly edit drivers/cam_cdm/cam_cdm_intf.c and friends.
         Then re-run add_EXPORT_SYMBOL_GPL to complete.
  INFO: export cam_cdm_intf_init_module in drivers/cam_cdm/cam_cdm_intf.c
  drivers/cam_cdm/cam_cdm_intf.c
  FAILED: could not complete task
  $ git diff
  . . .
  -int __init cam_cdm_intf_init_module(void)
  +int cam_cdm_intf_init_module(void)
  . . .
  +EXPORT_SYMBOL_GPL(cam_cdm_intf_init_module);
  . . .
  +EXPORT_SYMBOL_GPL(cam_cdm_intf_exit_module);
  . . .
  $ add_EXPORT_SYMBOL_GPL < build_output_file
  Reading symbols
  Finding symbols in the current tree
  Editing the files that contain the symbols
  INFO: cam_cdm_intf_exit_module found and already exported
  INFO: cam_cdm_intf_init_module found and already exported
  DONE
  WARNING: list of files in git different from files we expected to edit

Signed-off-by: Mark Salyzyn <salyzyn@google.com>
SideEffects: script was a one shot thing, now it allows continuation.
Test: manually create the condition and log since typically rare.
Change-Id: Ia5a719529317cc3679093a88f5eabdaf086233de